+++ apps/core/service.py
@@ -1,0 +1,53 @@
+from django.shortcuts import get_object_or_404, render, redirect
+from django.db import transaction
+from django.core.exceptions import ValidationError
+from apps.ecas.models import PuntoECA, Localidad
+from apps.users.models import Usuario
+
+
+class UserService:
+    @staticmethod
+    @transaction.atomic
+    def editar_perfil(request, id):
+        """
+        Vista para editar el perfil del gestor ECA.
+        """
+        # Obtener usuario o redirigir si no existe
+        try:
+            # usuario = Usuario.objects.get(id=id)
+            usuario = Usuario.objects.select_for_update().get(
+                id=id
+            )  # Bloqueo para evitar condiciones de carrera
+
+        except Usuario.DoesNotExist:
+            return redirect("punto:render_seccion", seccion="perfil")
+
+        # Actualizar campos básicos del usuario
+        usuario.nombres = request.POST.get("nombre", usuario.nombres)
+        usuario.apellidos = request.POST.get("apellido", usuario.apellidos)
+        usuario.email = request.POST.get("email", usuario.email)
+        usuario.celular = request.POST.get("telefono", usuario.celular)
+
+        # Manejo de la localidad como objeto
+        localidad_id = request.POST.get("localidad")
+        if localidad_id and localidad_id != str(
+            usuario.localidad.id if usuario.localidad else ""
+        ):
+            try:
+                usuario.localidad = Localidad.objects.get(id=localidad_id)
+            except Localidad.DoesNotExist:
+                pass  # Mantener la localidad actual si no existe la nueva
+
+        usuario.tipo_documento = request.POST.get(
+            "tipo_documento", usuario.tipo_documento
+        )
+        usuario.numero_documento = request.POST.get(
+            "numero_documento", usuario.numero_documento
+        )
+
+        try:
+            usuario.save()
+        except Exception:
+            pass  # Manejar errores silenciosamente por ahora
+
+        return usuario
