+++ apps/ecas/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from apps.ecas.views import render_seccion, editar_perfil
+from apps.ecas.views import render_seccion, editar_perfil_gestor
 
 app_name = "punto"
 
@@ -25,5 +25,5 @@ urlpatterns = [
     ),
     path("perfil/", render_seccion, {"seccion": "perfil"}, name="perfil"),
     path("resumen/", render_seccion, {"seccion": "resumen"}, name="resumen"),
-    path("editar-perfil/<str:id>/", editar_perfil, name="editar_perfil"),
+    path("editar-perfil/<str:id>/", editar_perfil_gestor, name="editar_perfil"),
 ]
